Shaded Maven plugin, first tests
I included the Shaded plugin from Maven to include dependencies in the built JAR file. This will make life a lot easier when testing code or including other dependencies or libraries, but is a quicker fix and might be too much with too many libraries. Users should build App.java with "mvn package" (the "Run" command in VSCode).